fix: Update created_at column in antrian table to use DATETIME instead of DATE
This commit updates the created_at column in the antrian table of the database schema. The column is changed from using the DATE data type to the DATETIME data type. This change allows for more precise timestamp recording, including the time component. It ensures accurate tracking of when an antrian (queue) is created.